Move `Stellar::Account` class to the `stellar-base` gem
