+++ sdk/lib/stellar/compat.rb
@@ -1,0 +1,3 @@
+class << Stellar::Account
+  deprecate deprecator: Stellar::SdkDeprecation, lookup: "Use `Stellar::Federation.lookup` method"
+end
