+++ sdk/spec/fixtures/vcr_cassettes/Stellar_Federation/_lookup/should_handle_404_request_when_performing_federation_lookup.yml
