+++ sdk/spec/fixtures/vcr_cassettes/Stellar_Federation/_lookup/should_handle_domains_that_are_not_federation_servers.yml
