+++ sdk/spec/fixtures/vcr_cassettes/Stellar_Federation/_lookup/should_peforms_federation_lookup.yml
